+++ src/RelaxedIT.EnergySaver/RelaxedIT.EnergySaver.psm1
@@ -8,7 +8,7 @@
     if (!(test-path -path $config ))
     {   $base = (Get-Module RelaxedIT.EnergySaver).ModuleBase
         copy-item -Path (join-path $base "EnergySaver.json") -Destination $config
-        Write-customLOG "[Initial]: copy default config: ""$config"""
+        Write-RelaxedIT "[Initial]: copy default config: ""$config"""
     }
 
     $processnames = (Get-ConfigfromJSON -config $config).id
@@ -24,14 +24,14 @@
             # Disable sleep mode
             powercfg -change -standby-timeout-ac 0
             powercfg -change -monitor-timeout-ac 10
-            Write-customLOG -logtext ($anyrunning.ProcessName -join ";")
-            Write-customLOG -logtext "is running. Sleep mode disabled."
+            Write-RelaxedIT -logtext ($anyrunning.ProcessName -join ";")
+            Write-RelaxedIT -logtext "is running. Sleep mode disabled."
         } else {
             $host.ui.RawUI.WindowTitle = "EnergyTimeout 20 Min"
             # Enable sleep mode after 20 minutes
             powercfg -change -standby-timeout-ac 20
             powercfg -change -monitor-timeout-ac 10
-            Write-customLOG -logtext "Neither ""$config"" is running. Sleep mode enabled after 20 minutes."
+            Write-RelaxedIT -logtext "Neither ""$config"" is running. Sleep mode enabled after 20 minutes."
         }
 
         # Wait for the interval before checking again